* make pubinternal classes public * move versionprefix and versionsuffix into Directory.Build.props * flatten namespaces, i.e. remove *.common * remove unnecessary csproj settings * README tweaks
