+++ samples/Samples.Schemas.Chat/Samples.Schemas.Chat.csproj
@@ -9,6 +9,7 @@
   <ItemGroup>
     <PackageReference Include="GraphQL.NewtonsoftJson" Version="$(GraphQLVersion)" />
   </ItemGroup>
+  
   <ItemGroup>
     <ProjectReference Include="..\..\src\Transports.Subscriptions.Abstractions\Transports.Subscriptions.Abstractions.csproj" />
   </ItemGroup>
