+++ samples/Samples.Server/GraphQLHttpMiddlewareWithLogs.cs
@@ -1,12 +1,10 @@
-using GraphQL.Server.Common;
-using GraphQL.Server.Transports.AspNetCore;
-using GraphQL.Server.Transports.AspNetCore.Common;
-using GraphQL.Types;
-using Microsoft.AspNetCore.Http;
-using Microsoft.Extensions.Logging;
 using System;
 using System.Threading;
 using System.Threading.Tasks;
+using GraphQL.Server.Transports.AspNetCore;
+using GraphQL.Types;
+using Microsoft.AspNetCore.Http;
+using Microsoft.Extensions.Logging;
 
 namespace GraphQL.Samples.Server
 {
