+++ src/Core/IGraphQLExecuter.cs
@@ -3,7 +3,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using GraphQL.Types;
 
-namespace GraphQL.Server.Internal
+namespace GraphQL.Server
 {
     /// <summary>
     /// GraphQL query, mutation and subscription executer
