+++ src/Transports.AspNetCore.NewtonsoftJson/GraphQLRequestDeserializer.cs
@@ -1,13 +1,11 @@
-using GraphQL.NewtonsoftJson;
-using GraphQL.Server.Common;
-using GraphQL.Server.Transports.AspNetCore.Common;
-using Microsoft.AspNetCore.Http;
-using Newtonsoft.Json;
 using System;
 using System.IO;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using GraphQL.NewtonsoftJson;
+using Microsoft.AspNetCore.Http;
+using Newtonsoft.Json;
 
 namespace GraphQL.Server.Transports.AspNetCore.NewtonsoftJson
 {
