+++ src/Transports.AspNetCore.NewtonsoftJson/InternalGraphQLRequest.cs
@@ -1,4 +1,3 @@
-using GraphQL.Server.Common;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
