+++ src/Transports.AspNetCore.SystemTextJson/GraphQLRequestDeserializer.cs
@@ -1,7 +1,3 @@
-using GraphQL.Server.Common;
-using GraphQL.Server.Transports.AspNetCore.Common;
-using GraphQL.SystemTextJson;
-using Microsoft.AspNetCore.Http;
 using System;
 using System.Buffers;
 using System.IO.Pipelines;
@@ -9,6 +5,8 @@ using System.Linq;
 using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
+using GraphQL.SystemTextJson;
+using Microsoft.AspNetCore.Http;
 
 namespace GraphQL.Server.Transports.AspNetCore.SystemTextJson
 {
