+++ src/Transports.AspNetCore.SystemTextJson/InternalGraphQLRequest.cs
@@ -1,4 +1,3 @@
-using GraphQL.Server.Common;
 using System.Collections.Generic;
 using System.Text.Json.Serialization;
