+++ src/Transports.Subscriptions.Abstractions/SubscriptionManager.cs
@@ -1,14 +1,13 @@
-using GraphQL.NewtonsoftJson;
-using GraphQL.Server.Internal;
-using GraphQL.Server.Transports.Subscriptions.Abstractions.Internal;
-using GraphQL.Subscription;
-using Microsoft.Extensions.Logging;
 using System;
 using System.Collections;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using GraphQL.NewtonsoftJson;
+using GraphQL.Server.Transports.Subscriptions.Abstractions.Internal;
+using GraphQL.Subscription;
+using Microsoft.Extensions.Logging;
 
 namespace GraphQL.Server.Transports.Subscriptions.Abstractions
 {
