+++ src/Transports.Subscriptions.WebSockets/Transports.Subscriptions.WebSockets.csproj
@@ -4,8 +4,6 @@
     <TargetFrameworks>netcoreapp3.1;netcoreapp3.0;netstandard2.0</TargetFrameworks>
     <AssemblyName>GraphQL.Server.Transports.WebSockets</AssemblyName>
     <RootNamespace>GraphQL.Server.Transports.WebSockets</RootNamespace>
-    <VersionPrefix>0.0.0</VersionPrefix>
-    <VersionSuffix>dev</VersionSuffix>
     <Description>WebSockets transport for subscriptions</Description>
     <PackageTags>GraphQL WebSockets subscriptions transport</PackageTags>
   </PropertyGroup>
