+++ src/Transports.Subscriptions.WebSockets/WebSocketConnectionFactory.cs
@@ -1,9 +1,8 @@
-using GraphQL.Server.Internal;
+using System.Collections.Generic;
+using System.Net.WebSockets;
 using GraphQL.Server.Transports.Subscriptions.Abstractions;
 using GraphQL.Types;
 using Microsoft.Extensions.Logging;
-using System.Collections.Generic;
-using System.Net.WebSockets;
 
 namespace GraphQL.Server.Transports.WebSockets
 {
