+++ tests/Samples.Server.Tests/ResponseTests.cs
@@ -1,12 +1,12 @@
 using System.Threading.Tasks;
 using Xunit;
-using GraphQL.Server.Common;
 using GraphQL;
 using Shouldly;
 using System.Net.Http;
 using System.Text;
 using System.Net;
 using System.Collections.Generic;
+using GraphQL.Server;
 
 #if NETCOREAPP2_2
 using GraphQL.NewtonsoftJson;
